Fixed a bug related to x64 compilation
size_t alias represents a ulong on a 64-bit systems so it’s not capable
with the int type

+++ source/artemisd/systems/entityprocessingsystem.d
@@ -19,7 +19,7 @@ abstract class EntityProcessingSystem : EntitySystem
 
     protected final override void processEntities(Bag!Entity entities) 
     {
-        for (int i = 0, s = entities.size(); s > i; i++) 
+        for (size_t i = 0, s = entities.size(); s > i; i++) 
         {
             process(entities.get(i));
         }
